Implement useful stuff for EraId
