+++ node/src/components/consensus/era_supervisor/era.rs
@@ -1,26 +1,19 @@
 use std::{
     collections::{BTreeMap, HashSet},
     env,
-    fmt::{self, Debug, Display, Formatter},
 };
 
 use datasize::DataSize;
 use itertools::Itertools;
 use once_cell::sync::Lazy;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
 use tracing::{debug, warn};
 
-use casper_types::{
-    bytesrepr::{self, FromBytes, ToBytes},
-    PublicKey, U512,
-};
+use casper_types::{PublicKey, U512};
 
 use crate::{
     components::consensus::{
         candidate_block::CandidateBlock, cl_context::ClContext,
         consensus_protocol::ConsensusProtocol, protocols::highway::HighwayProtocol,
-        ConsensusMessage,
     },
     types::{ProtoBlock, Timestamp},
 };
@@ -29,71 +22,6 @@ const CASPER_ENABLE_DETAILED_CONSENSUS_METRICS_ENV_VAR: &str =
     "CASPER_ENABLE_DETAILED_CONSENSUS_METRICS";
 static CASPER_ENABLE_DETAILED_CONSENSUS_METRICS: Lazy<bool> =
     Lazy::new(|| env::var(CASPER_ENABLE_DETAILED_CONSENSUS_METRICS_ENV_VAR).is_ok());
-
-#[derive(
-    DataSize,
-    Debug,
-    Clone,
-    Copy,
-    Hash,
-    PartialEq,
-    Eq,
-    PartialOrd,
-    Ord,
-    Serialize,
-    Deserialize,
-    JsonSchema,
-)]
-#[serde(deny_unknown_fields)]
-pub struct EraId(pub(crate) u64);
-
-impl EraId {
-    pub(crate) fn message(self, payload: Vec<u8>) -> ConsensusMessage {
-        ConsensusMessage::Protocol {
-            era_id: self,
-            payload,
-        }
-    }
-
-    pub(crate) fn successor(self) -> EraId {
-        EraId(self.0 + 1)
-    }
-
-    /// Returns the current era minus `x`, or `None` if that would be less than `0`.
-    pub(crate) fn checked_sub(&self, x: u64) -> Option<EraId> {
-        self.0.checked_sub(x).map(EraId)
-    }
-}
-
-impl Display for EraId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "era {}", self.0)
-    }
-}
-
-impl From<EraId> for u64 {
-    fn from(era_id: EraId) -> Self {
-        era_id.0
-    }
-}
-
-impl ToBytes for EraId {
-    fn to_bytes(&self) -> Result<Vec<u8>, bytesrepr::Error> {
-        self.0.to_bytes()
-    }
-
-    fn serialized_length(&self) -> usize {
-        self.0.serialized_length()
-    }
-}
-
-impl FromBytes for EraId {
-    fn from_bytes(bytes: &[u8]) -> Result<(Self, &[u8]), bytesrepr::Error> {
-        let (id_value, remainder) = u64::from_bytes(bytes)?;
-        let era_id = EraId(id_value);
-        Ok((era_id, remainder))
-    }
-}
 
 /// A candidate block waiting for validation and dependencies.
 #[derive(DataSize)]
@@ -324,20 +252,5 @@ where
             + slashed.estimate_heap_size()
             + accusations.estimate_heap_size()
             + validators.estimate_heap_size()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use rand::Rng;
-
-    use super::*;
-    use crate::testing::TestRng;
-
-    #[test]
-    fn bytesrepr_roundtrip() {
-        let mut rng = TestRng::new();
-        let era_id = EraId(rng.gen());
-        bytesrepr::test_serialization_roundtrip(&era_id);
     }
 }
